restructured application
application runs now. need to find a way to increment database now
