+++ my_app/__init__.py
@@ -1,6 +1,5 @@
-from flask import Flask
 
-app = Flask(__name__)
+
